# services/backend/poll/pkg/api/poll/v1/poll.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "poll/v1/poll.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "PollService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/polls": {
      "post": {
        "summary": "CreatePoll mints a poll and an opaque expiring link token (returned once).",
        "operationId": "PollService_CreatePoll",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreatePollResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreatePollRequest"
            }
          }
        ],
        "tags": [
          "PollService"
        ]
      }
    },
    "/v1/polls/token/{token}": {
      "get": {
        "summary": "GetPollByToken resolves the link token to its poll. Anonymous. Expired,\nrevoked, unknown or already-consumed tokens all return NotFound uniformly.",
        "operationId": "PollService_GetPollByToken",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetPollByTokenResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "PollService"
        ]
      }
    },
    "/v1/polls/token/{token}/responses": {
      "post": {
        "summary": "SubmitResponse records the Subject's answers. Anonymous, rate-limited, one\nresponse per poll; emits PollCompleted on success.",
        "operationId": "PollService_SubmitResponse",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SubmitResponseResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PollServiceSubmitResponseBody"
            }
          }
        ],
        "tags": [
          "PollService"
        ]
      }
    },
    "/v1/polls/{pollId}/responses": {
      "get": {
        "summary": "GetResponses returns a poll's responses. Owner-only (JWT subject must equal\nthe poll's owner_user_id).",
        "operationId": "PollService_GetResponses",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetResponsesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pollId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "PollService"
        ]
      }
    }
  },
  "definitions": {
    "PollServiceSubmitResponseBody": {
      "type": "object",
      "properties": {
        "answers": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Answer"
          }
        }
      },
      "title": "---- SubmitResponse (anonymous) ----"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Answer": {
      "type": "object",
      "properties": {
        "questionId": {
          "type": "string"
        },
        "text": {
          "type": "string",
          "title": "for TEXT questions"
        },
        "choiceIds": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "for CHOICE questions"
        }
      }
    },
    "v1CreatePollRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "questions": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Question"
          }
        },
        "surpriseRequestId": {
          "type": "string",
          "title": "optional"
        },
        "ttlSeconds": {
          "type": "string",
          "format": "int64",
          "title": "optional; server default if \u003c= 0"
        }
      },
      "title": "---- CreatePoll (owner) ----"
    },
    "v1CreatePollResponse": {
      "type": "object",
      "properties": {
        "pollId": {
          "type": "string"
        },
        "linkToken": {
          "type": "string",
          "title": "raw opaque token, returned ONCE"
        },
        "linkUrl": {
          "type": "string"
        },
        "expiresAt": {
          "type": "string",
          "title": "RFC3339"
        }
      }
    },
    "v1GetPollByTokenResponse": {
      "type": "object",
      "properties": {
        "pollId": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "questions": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Question"
          },
          "description": "Deliberately no owner_user_id / surprise_request_id: no owner data leaked."
        }
      }
    },
    "v1GetResponsesResponse": {
      "type": "object",
      "properties": {
        "responses": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Response"
          }
        }
      }
    },
    "v1Option": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "label": {
          "type": "string"
        }
      }
    },
    "v1Question": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "prompt": {
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/v1QuestionType"
        },
        "options": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Option"
          },
          "title": "for choice types"
        },
        "required": {
          "type": "boolean"
        }
      }
    },
    "v1QuestionType": {
      "type": "string",
      "enum": [
        "QUESTION_TYPE_UNSPECIFIED",
        "QUESTION_TYPE_TEXT",
        "QUESTION_TYPE_SINGLE_CHOICE",
        "QUESTION_TYPE_MULTI_CHOICE"
      ],
      "default": "QUESTION_TYPE_UNSPECIFIED",
      "title": "- QUESTION_TYPE_TEXT: free-text answer\n - QUESTION_TYPE_SINGLE_CHOICE: exactly one option id\n - QUESTION_TYPE_MULTI_CHOICE: one or more option ids"
    },
    "v1Response": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "answers": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Answer"
          }
        },
        "submittedAt": {
          "type": "string",
          "title": "RFC3339"
        }
      }
    },
    "v1SubmitResponseResponse": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean"
        }
      }
    }
  }
}